refactor(core): begin moving and refactoring withLoader, body, link
