--- packages/core/components/app/app.css
@@ -1,50 +1,0 @@
-.wrapper {
-  display: flex;
-  flex-direction: column;
-}
-
-.screenreaderOnly {
-  border: 0;
-  clip: rect(1px, 1px, 1px, 1px);
-  clip-path: inset(50%);
-  height: 1px;
-  margin: -1px;
-  overflow: hidden;
-  padding: 0;
-  position: absolute;
-  width: 1px;
-  word-wrap: normal;
-}
-
-.buttonReset {
-  appearance: none;
-  background-color: transparent;
-  border: none;
-  border-radius: 0;
-  font: inherit;
-  margin: 0;
-  padding: 0;
-}
-
-.skipLink {
-  align-items: center;
-  display: flex;
-  justify-content: center;
-  left: 0;
-  text-transform: uppercase;
-  top: 0;
-  z-index: 9999;
-}
-
-.skipLink:focus,
-.skipLink:active {
-  background-color: #FFFFFF;
-  border: 2px solid #000000;
-  clip: auto;
-  clip-path: none;
-  color: #000000;
-  height: 40px;
-  margin: 0;
-  padding: 0;
-  width: 200px;
-}
