--- packages/core/components/defaultErrorMessage/errorMessage.css
@@ -1,5 +1,0 @@
-.wrapper {
-  margin: 0 auto;
-  max-width: 980px;
-  text-align: center;
-}
