+++ packages/core/components/hoc/withAsync.js
@@ -1,5 +1,5 @@
 import universal from 'react-universal-component';
-import DefaultLoading from 'components/helpers/defaultLoading';
+import DefaultLoading from 'components/defaultLoading';
 
 const withAsync = (importer) => (
   universal(importer, {
