+++ packages/core/components/hoc/withLoader.test.js
@@ -3,7 +3,7 @@ import { mount } from 'enzyme';
 import { createStore } from 'redux';
 import { Provider } from 'react-redux';
 import defaultState from 'reducers/defaultState';
-import DefaultLoading from 'components/helpers/defaultLoading';
+import DefaultLoading from 'components/defaultLoading';
 import withLoader from './withLoader';
 
 it('should skip rendering wrapped component if loading', () => {
