--- packages/core/components/link/readme.md
@@ -1,5 +1,0 @@
-Link helper component. Responsible for handling history push for relative links.
-
-```js
-<Link to="#link" className="test">This is a sample link</Link>
-```
