+++ packages/core/hooks/useLoading.js
@@ -1,0 +1,7 @@
+import { useSelector } from 'react-redux';
+
+const useLoading = () => (
+  useSelector((state) => state.loading)
+);
+
+export default useLoading;
